Revert "Keep table of term replacements in SideCondition, use in term simplifier (#2315)"
This reverts commit c4278195b46c4c9e584582eff4d7e700afe661e2.
