+++ kore/src/Kore/Equation/Simplification.hs
@@ -37,6 +37,9 @@ import Kore.Internal.Pattern
     )
 import qualified Kore.Internal.Pattern as Pattern
 import qualified Kore.Internal.Predicate as Predicate
+import qualified Kore.Internal.SideCondition as SideCondition
+    ( top
+    )
 import qualified Kore.Internal.Substitution as Substitution
 import qualified Kore.Internal.TermLike as TermLike
 import qualified Kore.Step.Simplification.Pattern as Pattern
@@ -129,6 +132,6 @@ simplifyPattern
     :: (InternalVariable variable, MonadSimplify simplifier)
     => Pattern variable
     -> simplifier (OrPattern variable)
-simplifyPattern =
+simplifyPattern patt =
     Simplifier.localSimplifierAxioms (const mempty)
-    . Pattern.simplify
+    $ Pattern.simplify SideCondition.top patt
