+++ kore/src/Kore/Reachability/Claim.hs
@@ -2,8 +2,6 @@
 Copyright   : (c) Runtime Verification, 2019
 License     : NCSA
 -}
-{-# LANGUAGE Strict #-}
-
 module Kore.Reachability.Claim
     ( Claim (..)
     , ApplyResult (..)
@@ -136,6 +134,9 @@ import Kore.Step.Simplification.Data
     )
 import qualified Kore.Step.Simplification.Exists as Exists
 import qualified Kore.Step.Simplification.Not as Not
+import Kore.Step.Simplification.OrPattern
+    ( simplifyConditionsWithSmt
+    )
 import Kore.Step.Simplification.Pattern
     ( simplifyTopConfiguration
     )
@@ -526,9 +527,9 @@ checkImplicationWorker (ClaimPattern.refreshExistentials -> claimPattern) =
                 $ from $ makeCeilPredicate leftTerm
         let configs' = MultiOr.map (definedConfig <*) removal
         stuck <-
-            Logic.scatter configs'
-            >>= Pattern.simplify
-            >>= SMT.Evaluator.filterMultiOr
+            simplifyConditionsWithSmt sideCondition configs'
+            >>= Logic.scatter
+            >>= Pattern.simplify sideCondition
             >>= Logic.scatter
         pure (examine anyUnified stuck)
     & elseImplied
@@ -557,12 +558,12 @@ checkImplicationWorker (ClaimPattern.refreshExistentials -> claimPattern) =
             unified <-
                 mkIn sort leftTerm rightTerm
                 & Pattern.fromTermLike
-                & Pattern.simplify
+                & Pattern.simplify sideCondition
                 & (>>= Logic.scatter)
             didUnify
             removed <-
                 Pattern.andCondition unified rightCondition
-                & Pattern.simplify
+                & Pattern.simplify sideCondition
                 & (>>= Logic.scatter)
             Exists.makeEvaluate sideCondition existentials removed
                 >>= Logic.scatter
@@ -631,7 +632,7 @@ simplifyRightHandSide lensClaimPattern sideCondition =
     Lens.traverseOf (lensClaimPattern . field @"right") $ \dest ->
         OrPattern.observeAllT
         $ Logic.scatter dest
-        >>= Pattern.makeEvaluate sideCondition . Pattern.requireDefined
+        >>= Pattern.simplify sideCondition . Pattern.requireDefined
         >>= SMT.Evaluator.filterMultiOr
         >>= Logic.scatter
