+++ kore/src/Kore/Step/Simplification/Exists.hs
@@ -7,8 +7,6 @@ Maintainer  : virgil.serbanuta@runtimeverification.com
 Stability   : experimental
 Portability : portable
 -}
-{-# LANGUAGE Strict #-}
-
 module Kore.Step.Simplification.Exists
     ( simplify
     , makeEvaluate
@@ -80,7 +78,7 @@ import qualified Kore.Step.Simplification.AndPredicates as And
     ( simplifyEvaluatedMultiPredicate
     )
 import qualified Kore.Step.Simplification.Pattern as Pattern
-    ( makeEvaluate
+    ( simplify
     )
 import Kore.Step.Simplification.Simplify
 import qualified Kore.TopBottom as TopBottom
@@ -305,7 +303,7 @@ makeEvaluateBoundLeft sideCondition variable boundTerm normalized
                         $ Conditional.predicate normalized
                     }
         orPattern <-
-            lift $ Pattern.makeEvaluate sideCondition substituted
+            lift $ Pattern.simplify sideCondition substituted
         Logic.scatter (toList orPattern)
   where
     someVariableName = inject (variableName variable)
