+++ kore/src/Kore/Step/Simplification/Rule.hs
@@ -2,8 +2,6 @@
 Copyright   : (c) Runtime Verification, 2019
 License     : NCSA
 -}
-{-# LANGUAGE Strict #-}
-
 module Kore.Step.Simplification.Rule
     ( simplifyRulePattern
     , simplifyRewriteRule
@@ -24,6 +22,9 @@ import Kore.Internal.Predicate
     ( pattern PredicateTrue
     )
 import qualified Kore.Internal.Predicate as Predicate
+import qualified Kore.Internal.SideCondition as SideCondition
+    ( topTODO
+    )
 import qualified Kore.Internal.Substitution as Substitution
 import Kore.Internal.TermLike
     ( TermLike
@@ -143,4 +144,4 @@ simplifyPattern
     -> simplifier (OrPattern variable)
 simplifyPattern termLike =
     Simplifier.localSimplifierAxioms (const mempty)
-    $ Pattern.simplify (Pattern.fromTermLike termLike)
+    $ Pattern.simplify SideCondition.topTODO (Pattern.fromTermLike termLike)
