+++ kore/test/Test/Kore/Step/Axiom/Registry.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE Strict #-}
-
 module Test.Kore.Step.Axiom.Registry
     ( test_functionRegistry
     ) where
@@ -48,6 +46,9 @@ import qualified Kore.IndexedModule.MetadataToolsBuilder as MetadataTools
     ( build
     )
 import Kore.Internal.Pattern as Pattern
+import qualified Kore.Internal.SideCondition as SideCondition
+    ( top
+    )
 import Kore.Internal.Symbol as Symbol
 import Kore.Internal.TermLike
 import qualified Kore.Step.Axiom.Identifier as AxiomIdentifier
@@ -448,7 +449,7 @@ test_functionRegistry =
         let expect = mkApplySymbol sHead []
         simplified <-
             runSimplifier testEnv
-            $ Pattern.simplify
+            $ Pattern.simplify SideCondition.top
             $ makePattern $ mkApplySymbol gHead []
         let actual = Pattern.term $ head $ toList simplified
         assertEqual "" expect actual
@@ -456,7 +457,7 @@ test_functionRegistry =
         let expect = mkApplySymbol tHead []
         simplified <-
             runSimplifier testEnv
-            $ Pattern.simplify
+            $ Pattern.simplify SideCondition.top
             $ makePattern $ mkApplySymbol pHead []
         let actual = Pattern.term $ head $ toList simplified
         assertEqual "" expect actual
