+++ kore/test/Test/Kore/Step/Simplification/Integration.hs
@@ -57,7 +57,7 @@ import Kore.Step.Axiom.Registry
     ( mkEvaluatorRegistry
     )
 import qualified Kore.Step.Simplification.Pattern as Pattern
-    ( makeEvaluate
+    ( simplify
     )
 import Kore.Step.Simplification.Simplify
 
@@ -1405,7 +1405,7 @@ evaluateConditionalWithAxioms
     -> TestPattern
     -> IO OrTestPattern
 evaluateConditionalWithAxioms axioms sideCondition =
-    runSimplifierSMT env . Pattern.makeEvaluate sideCondition
+    runSimplifierSMT env . Pattern.simplify sideCondition
   where
     env = Mock.env { simplifierAxioms }
     simplifierAxioms :: BuiltinAndAxiomSimplifierMap
