+++ test/issue-2095/test-issue-2095.sh.out.golden
@@ -16502,9 +16502,7 @@
                         ),
                         /* Fl Fn D Sfc */
                         /* Inj: */ inj{SortInt{}, SortSimpleData{}}(
-                            /* Created: Kore.Step.RewriteStep.resetResultPattern */
-                            /* Fl Fn D Sfa */
-                            Var'QuesUnds'4:SortInt{}
+                            /* Fl Fn D Sfa */ Var'QuesUnds'4:SortInt{}
                         ),
                         /* Fl Fn D Sfa Cli */
                         /* Inj: */ inj{SortNumTypeName{}, SortTypeName{}}(
@@ -16557,9 +16555,7 @@
                                 ),
                                 /* Fl Fn D Sfc */
                                 /* Inj: */ inj{SortInt{}, SortSimpleData{}}(
-                                    /* Created: Kore.Step.RewriteStep.resetResultPattern */
-                                    /* Fl Fn D Sfa */
-                                    Var'QuesUnds'4:SortInt{}
+                                    /* Fl Fn D Sfa */ Var'QuesUnds'4:SortInt{}
                                 ),
                                 /* Fl Fn D Sfa Cli */
                                 /* Inj: */ inj{SortNumTypeName{}, SortTypeName{}}(
